created Durchstich for early detection of QR
The JNI code is a fucking mess. Somehow, it works, I am surpised to say the least. accessing the variables from another class in C++ was a major issue. I included the wechat_qrcode.hpp in nanodet.cpp, this allows me to share variables (I think). The namespaces also have to be taken into consideration. There is also sometimes the detection of false positives