implemented lock for canavs.
The first idea is rarely the best. Aquiring the canvas lock like this cannot be the best way. My approach is to continuously loop and just catch the exception, if the canvas is already locked (it is locked on the native side.) ... Maybe it's possible to use a boolean value that somehow exist in both memory pools. That's like using a volatile boolean for thread control of concurrent access, just not everything in in the JVM. This would (again) require communication over JNI. Might not be worth the complexity (and the increased possibility that an incorrect C++ implementaion.