Simplicity Proliferates. Finally, issue #34 fixed :)
This commit removes QrCodeActivity, so that we no longer have to manually transition between Activites and pass Arguments, which is very awkard and creates a bunch of boilerplate. Basically, I can put the CameraFragment inside the FragmentContainerView, and remove it very fast. To have a callback on the HostActivity when QR-Code related stuff is finished, I used a ViewModel called StateViewModel. This is a very simply ViewModel, only one value: State. I observe the MutableLiveData state in the NanodetncnnActivity so that I have a callback, for when ImageProcessing, Qr-Code and API is finished.